Change user key to integer, modify db scripts

+++ src/main/java/com/smarthouse/user/controller/UserController.java
@@ -26,7 +26,7 @@ public class UserController {
     }
 
     @GetMapping("/user/{id}")
-    public ResponseEntity<User> getUserById(@PathVariable("id") long id) {
+    public ResponseEntity<User> getUserById(@PathVariable("id") int id) {
         Optional<User> userOptional = userService.getById(id);
         return ResponseEntity.ok(userOptional.orElseThrow((Supplier<RuntimeException>) () ->
                 new ResourceNotFound(getNoResourceMessage("User", id))
